Move GP specific tests out of test_models.cc to test_gp.cc
